Revert "chore: Code style standardization and root directory cleanup"
This reverts commit a275db3fdc51131492ab92a35e863c83557b8776.

+++ src/GauntletCI.Tests/AuditLogTests.cs
@@ -108,10 +108,7 @@ public class AuditLogTests
         }
         finally
         {
-            if (File.Exists(tempFile))
-            {
-                File.Delete(tempFile);
-            }
+            if (File.Exists(tempFile)) File.Delete(tempFile);
         }
     }
 
@@ -142,10 +139,7 @@ public class AuditLogTests
         }
         finally
         {
-            if (File.Exists(tempFile))
-            {
-                File.Delete(tempFile);
-            }
+            if (File.Exists(tempFile)) File.Delete(tempFile);
         }
     }
 }
